Component node example
